pyproject.toml
#!/usr/bin/python3
#
# Copyright (C) 2022  Fernando Jurado-Lasso <ffjla@dtu.dk>

# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.

# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.

# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <https://www.gnu.org/licenses/>.


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = 'sdwsn_controller'
version = "1.0"
authors = [{ name = "Fernando Jurado-Lasso", email = "ffjla@dtu.dk" }]
description = "An open source implementation of the SDWSN controller for the ELISE project (https://github.com/fdojurado/ELISE)."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = [
    "Software Defined Wireless Sensor Networks",
    "SDWSN",
    "Software Defined Controller",
    "SDWSN controller",
    "WSN",
    "SDN",
    "SDN-IoT",
]
# dynamic = ["dependencies"]

dependencies = [
    'stable-baselines3[extra]',
    'matplotlib',
    'docker',
    'gym',
    'networkx',
    'pandas',
    'pyserial',
    'paho-mqtt',
    # 'plac',
    'pyfiglet',
    # 'pymongo',
    'python-daemon',
    'rich',
    'tomli'
]

[project.urls]
"Documentation" = "https://github.com/fdojurado/SDWSN-controller/wiki"
"Source" = "https://github.com/fdojurado/SDWSN-controller"
"Bug Tracker" = "https://github.com/fdojurado/SDWSN-controller/issues"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
